# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMLAB API",
        "description": "Dokumentasi REST API untuk Sistem Informasi Manajemen Laboratorium (SIMLAB) - Institut Teknologi Kalimantan",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pengajuan": {
            "get": {
                "tags": [
                    "Pengguna (Semua Role)"
                ],
                "summary": "Mengambil riwayat seluruh pengajuan milik user yang sedang login",
                "operationId": "a9c968769af572c460ab6cf609fcb6a4",
                "responses": {
                    "200": {
                        "description": "Daftar riwayat pengajuan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "tipe_pengajuan": {
                                                        "type": "string",
                                                        "example": "ruangan"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "diajukan"
                                                    },
                                                    "nomor_hp": {
                                                        "type": "string",
                                                        "example": "08123456789"
                                                    },
                                                    "judul_proyek": {
                                                        "type": "string",
                                                        "example": "Praktikum Fisika Dasar"
                                                    },
                                                    "dibuat_pada": {
                                                        "type": "string",
                                                        "example": "22 Apr 2026, 16:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pengguna (Semua Role)"
                ],
                "summary": "Membuat pengajuan baru (ruangan / alat / pengujian)",
                "description": "Endpoint tunggal untuk semua tipe pengajuan. Field yang diperlukan bervariasi tergantung `tipe_pengajuan`.",
                "operationId": "1ea50aedd65f42a4d4d705ab5462f410",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tipe_pengajuan",
                                    "nomor_hp",
                                    "tujuan_penggunaan"
                                ],
                                "properties": {
                                    "tipe_pengajuan": {
                                        "type": "string",
                                        "example": "ruangan",
                                        "enum": [
                                            "ruangan",
                                            "alat",
                                            "pengujian"
                                        ]
                                    },
                                    "nomor_hp": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "judul_proyek": {
                                        "type": "string",
                                        "example": "Praktikum Fisika Dasar",
                                        "nullable": true
                                    },
                                    "tujuan_penggunaan": {
                                        "type": "string",
                                        "example": "Digunakan untuk praktikum semester 2"
                                    },
                                    "dosen_pembimbing": {
                                        "type": "string",
                                        "example": "Dr. Ahmad Fauzi",
                                        "nullable": true
                                    },
                                    "email_dosen": {
                                        "type": "string",
                                        "example": "fauzi@lecturer.itk.ac.id",
                                        "nullable": true
                                    },
                                    "ruangan_id": {
                                        "description": "Wajib jika tipe = ruangan",
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "tanggal_mulai": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-05-01",
                                        "nullable": true
                                    },
                                    "tanggal_selesai": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-05-01",
                                        "nullable": true
                                    },
                                    "waktu_mulai": {
                                        "type": "string",
                                        "example": "08:00",
                                        "nullable": true
                                    },
                                    "waktu_selesai": {
                                        "type": "string",
                                        "example": "12:00",
                                        "nullable": true
                                    },
                                    "jumlah_pengguna": {
                                        "type": "integer",
                                        "example": 20,
                                        "nullable": true
                                    },
                                    "alat_id": {
                                        "description": "Wajib jika tipe = alat",
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    },
                                    "jumlah_dipinjam": {
                                        "type": "integer",
                                        "example": 3,
                                        "nullable": true
                                    },
                                    "tanggal_mulai_alat": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-05-01",
                                        "nullable": true
                                    },
                                    "tanggal_selesai_alat": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-05-03",
                                        "nullable": true
                                    },
                                    "keperluan_spesifik": {
                                        "type": "string",
                                        "example": "Untuk pengukuran tegangan rangkaian",
                                        "nullable": true
                                    },
                                    "jenis_pengujian_id": {
                                        "description": "Wajib jika tipe = pengujian",
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "nama_sampel": {
                                        "type": "string",
                                        "example": "Beton campuran A",
                                        "nullable": true
                                    },
                                    "jumlah_sampel": {
                                        "type": "integer",
                                        "example": 5,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pengajuan berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "tipe_pengajuan": {
                                                    "type": "string",
                                                    "example": "ruangan"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "diajukan"
                                                },
                                                "dibuat_pada": {
                                                    "type": "string",
                                                    "example": "22 Apr 2026, 16:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/pengajuan/{id}": {
            "get": {
                "tags": [
                    "Pengguna (Semua Role)"
                ],
                "summary": "Mengambil detail satu pengajuan berdasarkan ID",
                "operationId": "d3365b4ee3fbdd9f8bb942dc1e89c660",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail pengajuan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "tipe_pengajuan": {
                                                    "type": "string",
                                                    "example": "ruangan"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "diajukan"
                                                },
                                                "judul_proyek": {
                                                    "type": "string",
                                                    "example": "Praktikum Fisika Dasar"
                                                },
                                                "tujuan_penggunaan": {
                                                    "type": "string",
                                                    "example": "Digunakan untuk praktikum semester 2"
                                                },
                                                "dibuat_pada": {
                                                    "type": "string",
                                                    "example": "22 Apr 2026, 16:00"
                                                },
                                                "detail_ruangan": {
                                                    "properties": {
                                                        "ruangan": {
                                                            "type": "string",
                                                            "example": "Lab. Fisika Dasar A (R202)"
                                                        },
                                                        "tanggal_mulai": {
                                                            "type": "string",
                                                            "example": "2026-05-01"
                                                        },
                                                        "tanggal_selesai": {
                                                            "type": "string",
                                                            "example": "2026-05-01"
                                                        },
                                                        "waktu_mulai": {
                                                            "type": "string",
                                                            "example": "08:00:00"
                                                        },
                                                        "waktu_selesai": {
                                                            "type": "string",
                                                            "example": "12:00:00"
                                                        }
                                                    },
                                                    "type": "object",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Anda tidak memiliki akses ke pengajuan ini"
                    },
                    "404": {
                        "description": "Pengajuan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
...
    },
...
}